Lightning

+++ userbot/plugins/tagall.py
@@ -1,29 +1,16 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2
-import asyncio
-from telethon import events
 from telethon.tl.types import ChannelParticipantsAdmins
+
 from userbot.utils import admin_cmd
 
-
-@borg.on(admin_cmd(pattern=r"tagall", outgoing=True))
-async def _(event):
-    if event.fwd_from:
-        return
-    mentions = "Hello"
-    chat = await event.get_input_chat()
-    async for x in borg.iter_participants(chat, 100):
-        mentions += f" \n [{x.first_name}](tg://user?id={x.id})"
-    await event.reply(mentions)
-    await event.delete()
+# Added to userbot by @its_xditya
 
 
 @borg.on(admin_cmd(pattern=r"administrator", outgoing=True))
+@borg.on(sudo_cmd(pattern=r"administrator", allow_sudo=True))
 async def _(event):
     if event.fwd_from:
         return
-    mentions = "Administrators : "
+    mentions = "Administrators in the chat : "
     chat = await event.get_input_chat()
     async for x in borg.iter_participants(chat, filter=ChannelParticipantsAdmins):
         mentions += f" \n [{x.first_name}](tg://user?id={x.id})"
@@ -33,4 +20,20 @@ async def _(event):
         await reply_message.reply(mentions)
     else:
         await event.reply(mentions)
+    await event.delete()
+
+
+# Added to userbot by @its_xditya
+
+
+@borg.on(admin_cmd(pattern=r"tagall", outgoing=True))
+@borg.on(sudo_cmd(pattern=r"tagall", allow_sudo=True))
+async def _(event):
+    if event.fwd_from:
+        return
+    mentions = "Hey there!"
+    chat = await event.get_input_chat()
+    async for x in borg.iter_participants(chat, 100):
+        mentions += f" \n [{x.first_name}](tg://user?id={x.id})"
+    await event.reply(mentions)
     await event.delete()
